refactor(bottomnav): convert BottomNav nav parent itself; collapse its when block
BottomNavScreen moves to bottomnav/impl/ui. BottomNavBloc extends
BlocScreen; BottomNavBloc.Child also extends BlocScreen so the
BottomNavContentContainer's when block collapses to a single
`created.instance.Content()`.

BottomNavBlocImpl overrides Content() to call BottomNavigationScreen.
RootBloc.Child.BottomNavigation delegates via BlocScreen by bloc;
RootScreen renders via child.Content().

BottomNavBloc.Tab.getLabel()/.getIcon() go back to `internal` —
both BottomNavScreen and BottomNavOrderScreen now live in the same
bottomnav/impl/ui package, so the public visibility added during
the BottomNavOrder commit is no longer needed.
