+++ client/root/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/root/RootBloc.kt
@@ -25,7 +25,7 @@ interface RootBloc : BackHandlerOwner, BackClickBloc {
     fun handleSharedUrl(url: String)
 
     sealed class Child {
-        data class BottomNavigation(val bloc: BottomNavBloc) : Child()
+        data class BottomNavigation(val bloc: BottomNavBloc) : Child(), BlocScreen by bloc
 
         data class GroceryDetail(val bloc: GroceryDetailBloc) : Child(), BlocScreen by bloc
